+++ hooks/end_of_file.py
@@ -1,11 +1,25 @@
+"""
+# end-of-file
+#
+# Ensures that a file is either empty, or ends with exactly one newline.
+# Optionally rewrites offending files in place when --fix=yes is given.
+#
+# Copyright ©2021-2026 Marcin Orlowski <mail [@] MarcinOrlowski.com>
+# https://github.com/MarcinOrlowski/pre-commit-hooks/
+#
+# Test invocation:
+#   pre-commit try-repo . end-of-file --verbose --all-files
+"""
+
 import argparse
 import os
+import sys
 from typing import IO
 from typing import Optional
 from typing import Sequence
 
 
-def fix_file(args, file_obj: IO[bytes]) -> int:
+def fix_file(args: argparse.Namespace, file_obj: IO[bytes]) -> int:
     """
     Test for newline at end of file Empty files will throw IOError here
 
@@ -16,7 +30,7 @@ def fix_file(args, file_obj: IO[bytes]) -> int:
     except OSError:
         return 0
 
-    last_character = file_obj.read(1)
+    last_character: bytes = file_obj.read(1)
     # last_character will be '' for an empty file
     if last_character not in {b'\n', b'\r'} and last_character != b'':
         if args.fix:
@@ -41,8 +55,8 @@ def fix_file(args, file_obj: IO[bytes]) -> int:
 
     # Our current position is at the end of the file just before any amount of
     # newlines.  If we find extraneous newlines, then backtrack and trim them.
-    position = file_obj.tell()
-    remaining = file_obj.read()
+    position: int = file_obj.tell()
+    remaining: bytes = file_obj.read()
     for sequence in (b'\n', b'\r\n', b'\r'):
         if remaining == sequence:
             return 0
@@ -56,24 +70,24 @@ def fix_file(args, file_obj: IO[bytes]) -> int:
 
 
 def main(argv: Optional[Sequence[str]] = None) -> int:
-    parser = argparse.ArgumentParser()
+    parser: argparse.ArgumentParser = argparse.ArgumentParser()
     parser.add_argument('--fix', action = 'store', dest = 'fix', default = "no",
                         help = 'Corrects invalid files in-place. Supported values: "yes", "no" (default).')
     parser.add_argument('filenames', nargs = '*', help = 'Filenames to fix')
-    args = parser.parse_args(argv)
+    args: argparse.Namespace = parser.parse_args(argv)
 
     if args.fix not in ["yes", "no"]:
         print(f'Invalid --fix value: {args.fix}')
         return 10
-    args.fix = args.fix == "yes"
+    args.fix: bool = args.fix == "yes"
 
 
-    retv = 0
+    retv: int = 0
 
     for filename in args.filenames:
         # Read as binary so we can read byte-by-byte
         with open(filename, 'rb+') as file_obj:
-            ret_for_file = fix_file(args, file_obj)
+            ret_for_file: int = fix_file(args, file_obj)
             if ret_for_file:
                 print(f'Fixing {filename}')
             retv |= ret_for_file
@@ -82,4 +96,4 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
 
 
 if __name__ == '__main__':
-    exit(main())
+    sys.exit(main(sys.argv[1:]))